Update header components
