+++ frontend/app/common/header/components/MobileCategories.tsx
